Remove unused flag bit
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@341 42af7a65-404d-4744-a932-0658087f49c3

+++ net/uip/uip.c
@@ -175,7 +175,7 @@ uint16 uip_urglen, uip_surglen;
 uint16 uip_len, uip_slen;        /* The uip_len is either 8 or 16 bits, depending
                                   * on the maximum packet size. */
 
-uint16 uip_flags;                /* The uip_flags variable is used for communication
+uint8  uip_flags;                /* The uip_flags variable is used for communication
                                   * between the TCP/IP stack and the application
                                   * program. */
 struct uip_conn *uip_conn;       /* uip_conn always points to the current connection. */
